Fix workout recording crash

+++ app/src/main/java/de/tadris/fitness/util/AltitudeCorrection.java
@@ -1,3 +1,22 @@
+/*
+ * Copyright (c) 2020 Jannis Scheibe <jannis@tadris.de>
+ *
+ * This file is part of FitoTrack
+ *
+ * FitoTrack is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     FitoTrack is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package de.tadris.fitness.util;
 
 import android.content.Context;
@@ -43,7 +62,7 @@ public class AltitudeCorrection {
             String[] data = line.split(",");
             int lat = Integer.parseInt(data[0]);
             int lon = Integer.parseInt(data[1]);
-            double offset = Integer.parseInt(data[2]);
+            double offset = Double.parseDouble(data[2]);
             if (lat == this.latitude && lon == this.longitude) {
                 this.offset = offset;
                 return;
